Printing large-scale chirped gratings
Generates chirped sawteeth SLM images and uses a modified version of Julian's printing code to print them. 

-chirped-SLMimage.py
Chirp is defined by a starting and ending xmax and number of columns (SLM_images) to generate. Inputs are given by editing the code. A smoothly chirped sawtooth waveform with pixel width equal to the SLM width times the number of columns is created. Then individual SLM images are cropped out and saved with a number preceding their name denoting the column they represent. The right most column is printed first and denoted with a 1. 

-chirped_printer.py
Julian's printing code is quickly modified to display a new image on the SLM at the start of each print line (the stitching code is unchanged). The code now requires you to edit it to define the 'basic image name,' which is the name of the images generated by the chirped SLMimage script without the preceding number. When selecting print be sure to select the folder where the chirped SLM images are saved. The GUI will still prompt you to choose a color but your choice is irrelevant. 

-SerialMotorControl_Active.py 
unchanged, placed here because it is a dependency of chirped_printer.py